requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.1
openai>=1.1.0
pandas>=2.0.0
numpy>=1.20.0
streamlit>=1.24.0
weaviate-client>=4.4.0
sentence-transformers>=2.2.2
tavily-python>=0.2.0
tqdm>=4.65.0
python-dotenv>=1.0.0
openpyxl>=3.1.0 